Start builting out the basic config structure
To do this I needed to make some changes to how we have TypeScript
configured. I think we should start by using project references so we
get clean types in all of our projects and have builds with a sane
dependency graph from the getgo. I think the way I have these tsconfig
files set up we should be able to get types for everything, but only the
things in src will be built to dist for packaging.