Support all of the js and ts extensions for the config file
This will make it easier for people to use a Happo config file in their
repo, regardless of the extension.

I added some tests and a test util. While working on this I decided to
change our typescript set up to use a solution file instead of nested
tsconfig files all over the place. This should make things simpler to
work with.