Resolve TypeScript error when assigning JSDOM to globalThis.window in test
TypeScript doesn't like that we are overriding globalThis.window with
JSDOM's window. Maybe there's a better solution to this, but for now I
am just going to tell TypeScript that we expect an error here and it
will be okay.

I resolved the second TypeScript error here by restoring the original
window value, whatever it may be.