Make tsconfig.browser.json more generalized
By setting the rootDir and outDir to be the same as we use in
tsconfig.dist.json, we can get the types to output in the same place
identically. This means we can include more directories for browser code
and the types will all just be in the right spot!