Replace crypto-js with local md5 implementation
I want to avoid bringing in depedencies if we can help it. This
dependency is unmaintained and is close to 500 KiB. MD5 isn't changing
ever, so we can drop in a good implementation and be done with it.

I found one here:

https://github.com/cotag/ts-md5/blob/master/src/md5.ts

I copied this in and updated it a bit so it would better match our
coding style and resolve some TypeScript errors with our configuration.