Add takeDOMSnapshot playwright test
This test is verifying some important things about the takeDOMSnapshot
function, things that are hard to test using jsdom (which the other
takeDOMSnapshot.test.ts file is using).

In order to use page.addInitScript, I needed all the js for
takeDOMSnapshot to be in one bundle. I spent some time getting js
modules to work in the browser but ran into several issues along the
way, including importing npm modules in the js (parse-srcset). And while
struggling with this I realized that even if I can get the local tests
to work with js modules, we still need a single bundle for the
happo-playwright module. Expecting people to serve local files coming
from us will be a major pain so the bundle will be vital there.

Cursor suggested esbundle and it was very easy to set up. I used
browserify in the past but I believe esbundle is more modern.