Output browser .js code to dist
This code will be run in browsers, so we need something that browsers
can parse, which means we need to output .js files from our .ts files
somewhere.

`emitDeclarationOnly: false` is not compatible with
`allowImportingTsExtensions: true`, so I had to change that at the same
time and update our imports to use `.js` extensions (which I find a
little strange, but that's how it works). To make the node test runner
able to resolve these imports, I ended up bringing in tsx.