Start moving over controller and dependencies from happo-e2e
This ended up being a beast. This package is depending on happo.io and
there was a bit of a cascade moving things over.

The build is broken right now and the tests are not passing, but I
wanted to get this in as a milestone. Subsequent commits can start
fixing issues.