Add tsconfig.playwright.json
This code was not included in any tsconfig file, which means that it was
not being type checked when running pnpm build. We don't want to share
code between playwright tests and node tests, so I made a new tsconfig
file.

While I was at it I moved some config around so we have some nicer
defaults in our base config.