Make RemoteBrowserTarget.execute build successfully
This isn't fixing all the build issues we have, but it gets us closer to
a clean setup. To simplify some things I decided to scrap sync execution
(which we are going to discontinue anyway).

I struggled with having no `viewport` definition on the BaseTarget
interface. To make things easier to work with, I moved it from the
DesktopBrowserTarget to the BaseTarget. We can revisit this if we want
to be more strict about not having a viewport on the mobile targets.