Bring in undici to support proxied fetch
The fetch that ships with Node does not expose a way to do a proxied
request. The current advice is to use undici, which is developed by the
NodeJS team and is what NodeJS uses internally, to make these requests.

More info:

https://github.com/nodejs/node/issues/43187

Their best practices documentation shows doing this in a different way,
but that requires changing from fetch to client.request, which is a
bigger change than I care to take on at the moment:

https://undici-docs.vramana.dev/docs/best-practices/proxy