Remove unused maxTries type from makeRequest
Some of the callsites needed to be updated at the same time. I bet this
is an actual bug we have right now with the packages that people are
using, and I think it may explain a recent mystery where we saw on of
our jobs fail during a deploy and it looked like the request was not
retried.