Make public Config interface free from internal properties
I had initially made Target contain a __dynamic flag, which we use when
creating dynamic targets in a Cypress/Playwright test suite. Instead of
polluting the public interface I'm adding an internal one that we can
depend on in the code, while still keeping the public interface public
and clean.