Remove support for HTTP_PROXY
This was brought over when we migrated old code into this repo and after
some investigation we think that we don't actually need to support this
anymore. Removing to simplify things for now--we can bring it back later
if it becomes important.

I had previously added the undici dependency to support this, so with
this change I am removing it.