Add Happo API tokens to tests
To make the Playwright tests run, we need a HAPPO_API_KEY and a
HAPPO_API_SECRET variable. To make this work locally I opted for the new
`--env-file` support from Node (using `--env-file-if-exists` so that
things don't break in CI).

I added a target in our happo.config.ts config so that it would match
the ones in happo-playwright.