Stop using `dev-` and __LATEST__ for report shas
Instead resolve the current sha from the git repo. If there are
outstanding changes, include them when computing the sha. This will
ensure that there's a unique sha when developing locally.

__LATEST__ was a shortcut into making e2e runs work on the first try. It
is recognized on the server and resolves to (I believe) the most recent
approved report in a comparison (by date). It's hard to predict what the
baseline will be in this scenario so it's better to let the Happo server
decide on compare time. It will
- Find the baseline from the sha1/beforeSha and the fallback shas
- If none is found, create an empty report for the beforeSha and compare
  against it.