Set up window.happo namespace for happo functions
I want the type checking in our screenshot function to be as good as
possible, so to get this all wired up in a nice way I am expanding the
API of what the browser bundle does a bit to include another function.
To help avoid conflicts, I am moving everything into a happo object.