Update takeDOMSnapshot to allow any Element type
We want to support more than just HTMLElement.

Though now I am seeing that the code in side of the page.evaluate
callback is not type checked as I expected, so this assertion is not
really necessary to ensure a passing typescript build. This is
unfortunate because it makes the code less safe, but I think the runtime
checks here are still probably useful to keep around.

I think maybe this should be moved inside of takeDOMSnapshot which could
be updated to accept a Node instead of an Element, but that change will
take a bit more work than what I want to tackle at the moment.