Pass along --config to wrapped e2e command
An issue that I've seen a few times with happo-e2e is that the config
loaded in the wrapper (`happo-e2e --`) is different from the one used
in the wrapped command (e.g. `playwright test`). To help smooth this
out, I'm passing along an environment variable, HAPPO_CONFIG_FILE, for
the path to the config used in the wrapping e2e command.