Refactor loadConfig env var test
We want to still tmpfs.mock in this test with a config file to
demonstrate that the env var takes precedence over the file finding.

While I was at it I also moved the process.env resetting more globally.
I wanted to do this in some global test setup and teardown, but that
isn't available until Node 24:
https://nodejs.org/api/test.html#global-setup-and-teardown

When I did this I realized that the way we were resetting process.env
somewhere else was not actually working correctly, so I updated that as
well.