Make window.happo and all properties optional
We don't have super strong guarantees that this object will be defined
and what will be on it. So I think we should play it safe and make
everything possibly undefined so we have to add checks before using it.

This necessitated moving the assertElement call into takeDOMSnapshot,
but that's okay since that's where I wanted it anyway.