Move over Storybook test app from happo-plugin-storybook
We are using this app for testing the storybook plugin. As part of
moving things over, I did the following:
- Converted js files to ts files
- Switched away from jsx to pure ts
- Removed two imports used for testing that are no longer needed:
  react-tether and @mui/material.

For react-tether, this is the commit that added it:
https://github.com/happo/happo-plugin-storybook/commit/c0c47aa5f7ce22ffc1de93bab277598861550367

In that commit I refer to a #__snapshot-box element. This element is no
longer part of the library, and I can't fully understand why the tether
library is necessary. We already have the React portal case test which
should serve the same purpose.

@mui/material was added as part of
https://github.com/happo/happo-plugin-storybook/pull/74 and it has a
commit in there to make beforeScreenshot async. We have other tests
verifying async behavior and this one is redundant.