Start storybook server for storybook playwright test
One alternative is to add a webServer config to the Playwright config,
but then it will need to build storybook before running any playwright
tests, which is not ideal for us. Since we only have one test where we
need this, I decided to scope it to just that one test.