Make `projects` top-level config property
After our (Joe+Henric) discussion on the bridge in Minneapolis, we've
decided to have `projects` be a map in the config. Each project can have
its own integration configured, and you specify the project in the CLI.

We're doing this change now so that when we start implementing the main
happo CLI we don't have to adapt too much after the feature is there.

Projects are now required to have a name through the map. Since we use
the name as the key, naming becomes a little easier since it would
suddenly feel a little odd to have long complicated names. Before when
it was just a string it was easier to make it long. Also with the map we
are guaranteed uniqueness.

I haven't made much use of the `integrationType` yet, but the idea is
that this property decides what path we take through the CLI. It might
also make some configuration/setup easier as we can assume more things
based on the type.