Fix types for Cypress folder
There are two main files here that are used slightly different:
- cypress/index.ts is used in the Cypress browser
- cypress/task.ts is used in the Cypress runner in Node.js

I thought about splitting up the ts configs for the two files but
decided it wasn't worth the complexity. So for now they are represented
in the same config.

I've made the decision to cut support for local screenshots for now. I
believe there are better ways to handle local screenshots (e.g. the
happo gradle plugin).