Add simple Cypress test setup
We will use this to test the happo cypress plugin. For now there's
nothing related to happo in here though, that will come next.

I'm choosing not to copy over the existing tests in happo-cypress as
those are somewhat complicated, and they mostly test the same
functionality that we are testing in the e2e folder.