Split up Playwright and Cypress Happo configs
Right now we are creating reports for Cypress and Playwright in the same
project in our Happo account. To split them up I'm making two different
happo.config.ts files, where each of them inherits/imports the base
config.

While I was at it I fixed an issue with using a relative path to the
config file. We'll absolutize it as soon as possible in the CLI so that
we are sure that it will be available across the board.