Remove e2e parts from CLI
Instead of having an e2e subcommand, we can rely on the dash-dash part
to trigger the right behavior.

The port used by the Happo wrapper doesn't have to be specified -- we
can get one automatically and use it through the http server start. I
unified the test server we use in some tests with the wrapping test
server so that there's one implementation.

Next I'll work on removing the `--allowFailures` option. We can specify
that in configuration instead which feels more useful.