Refactor makeHappoAPIRequest to take path and Happo config
This makes usage of this a lot nicer, since we can just pass it the
config and the API path.

We have one place where we receive a URL from the API and then pass it
through, so we still need to support passing a URL to the function. At
some point it might be nice to refactor that API endpoint to return a
path as well, which would allow us to simplify this function a bit and
make the types work more nicely. For now this seems okay though.

TypeScript made this refactor relatively painless.