Refactor environment test a little
I wanted to see why this test was slower than the rest. It turned out to
be related to setting up the git repo. I took a small swing at making
that faster, and added a few options that chat suggested. I'm not sure
it will really make much of a difference.

While I was doing this I ran into an error with the git operation, but
the error message did not show the information I needed because it was
in stdout and not stderr. I also spotted an opportunity to clean up the
env test a bit by adding a helper function for creating and committing a
new file.