Fix up some types for Happo static
We had some repetition of some types that we define more generally and I
wanted to get everything on the same page a little better. This is my
attempt at that.

While I was doing this, I spotted a bug in our waitForContent type.
Looking at the documentation, it seems that this should be a string and
not a function that returns a boolean.

I think there is more generalization and standardization we could do
here, but this seems like a step in the right direction.