Add Happo Orchestrate job
This will make it so we have one status check instead of a bunch.

I pulled this job out of our main Happo codebase and generalized it to
be a GitHub action. I think this might be useful to expose as part of
our public API once we get a full release here.