Rewrite test to use TypeScript directly
This test depended on the existence of files output from the build, but
this means that we need to build before running the tests which isn't
very great. Instead, we can use TypeScript directly here to test the
same thing.

I did this in a way that reads the tsconfig files and determines which
files will be emitted where. This is much faster than an alternative
approach where we actually run the build inside the test and store where
the files would be emitted without actually writing the files, but at
the risk of being incorrect. I think it is fine for this test though and
it is super fast.

Maybe we should move some of these tsconfig utility functions from this
test and the tsconfigs test into our test-utils directory so they can be
more easily reused.