Build .ts to .js and specify files list in package.json
The files field will ensure that we are only distributing the code that
people need.

And I decided to build the files to .js files since that's the advice
that Node docs gives:

> Node does not strip types in node_modules because it can cause
> significant performance issues for the official TypeScript compiler
> (tsc) and parts of VS Code, so the TypeScript maintainers would like
> to discourage people publishing raw TypeScript, at least for now.

https://nodejs.org/en/learn/typescript/publishing-a-ts-package

I decided to use esbuild for this since we are already using it in the
repo.

I am using the bundle option which will produce a single file for each
of the entry points/output files, except for where there is overlap
among the entrypoints and code split points (i.e. dynamic imports). This
will improve performance since node won't have to go resolving all of
our imports all of the time. I took this opportunity to convert some
imports to dynamic imports for the CLI, which will improve startup speed
of the CLI a bit.

This might also open the door for us to use JSX in our Storybook plugin
again.

I'm not entirely sure I got the configs correct for everything, in
particular the browser builds--I think we maybe want to use iife for
those.

I ended up needing to move the types directory into the dist directory
so the types would be right next to the built code they are providing
types for. This allows local imports pointing at the dist directory to
work.