Replace old happo with new happo
We have been working on a new version of the package with lots of
breaking changes. We started work in a different repo but decided to
move it into this one. This commit completes this move by deleting all
of the old files and moving everything in the new directory that was
brought in using git filter-repo down to the top level.