Disable eslint rule in md5 implementation
This started failing linting after the unicorn plugin updated. I
understand that fromCodePoint is better for handling more of unicode,
but I'm not sure if that is important here and fromCharCode is much
faster, which might be important. For now I'm just going to carry on
with what we have.