Fix static package validation
There were some adjustments that needed to be made to this now that we
generate the iframe.html if it does not exist.

Since we do this as part of the command, it makes it difficult to
actually test the validation logic in an integration style test, since
it should always exist, so I ended up just testing the two most likely
possibilities--either the file exists already or it does not and it gets
generated.