Use generated afterSha when local changes are stashed
I ran into a confusing thing when working on adding Happo pages tests to
the project. When I pushed my first pull request the comparison for the
new pages project found a before report associated with the main branch.

https://github.com/happo/happo/pull/263
https://happo.io/a/768/p/2454/compare/ddb309c460d0904724436f2e8699255189e345f2/bca61033ff26e952bd6b9cc51990451ddf7f0a2c?jobId=1948528

It turns out that I was running happo locally and at some point I ran a
report while I had all local changes added to the git index. We weren't
taking this state into account which made the afterSha resolve to the
current HEAD, which was a commit on the main branch.