Replace HAPPO_DOWNLOAD_ALL with config option
Instead of having this magical env variable, we instead allow people to
configure their cypress/playwright integrations with a boolean
downloadAllAssets.

While looking at tests, I decided to ditch the custom server and use
startTestServer instead. But I had to make the port static since we
assert some hashes in the package. With a dynamic port, it was hard to
know what the hashes would be.