Remove --githubBase CLI flag
I realized that this was only used when determining the link URL for
Travis CI. I don't think this is very useful, so I decided to simplify
things. The link can still be passed in directly via the `--link` flag.
This simplifies the API of our tool a bit, which will make it easier for
everyone to use, and for us to maintain.

I also noticed that there is a CIRCLE_PULL_REQUEST env var we can use
here, which matches what we do for other CI providers.

At some point it might be nice to refactor this env var soup to
something with a bit more struture, where we detect the CI env somehow
and then hand things off to a helper for that specific CI provider to
resolve these things.