Stop using basename when zipping in-memory content
When I replaced archiver with fflate, I made a mistake here and decided
to use basename to remove the path from these in-memory files. This is
because the test for createAssetPackage was checking the entry.name
which was just the basename. But the actual file was in fact stored in
the directory. This broke external images in Playwright and Cypress
tests.

The fix is to stop using basename, and update the assertions.