Don't use local changes when resolving afterSha in CI
In case the CI environment has one or more changes in local files
(possibly not part of the git index or the app itself) we shouldn't
include those changes in the sha resolution. Locally we want that to
ensure we get unique shas for unique changes but we need to treat CI as
a pristine environment (even if it isn't).