Throw if apiKey or apiSecret are missing in config file
There's probably more validation we could do here, but this is a good
step in the right direction that should help people out if they happen
to run into this issue. This will prevent them from running into more
obscure errors later in the execution, such as "Zero-length key is not
supported".