Improve package.json metadata for storybook integration catalog
Documentation on this can be found here:

  https://storybook.js.org/docs/addons/integration-catalog#addon-metadata

For supported frameworks, while we haven't tested this with everything
in the list, I'm not aware of anything that would prevent it from
working with any of these frameworks, so I decided to just use the
entire list.

One requirement listed for addons, is this:

> preset.js file written as an ES5 module at the root level

While we have a preset.js file, it is not at the root level of the
package. I am hoping that it will be okay though. We'll see!