Set device scale factor to 2, allow for configuration
I have been investigating some flake caused by images with border-radius
and background colors. I determined that it can be resolved by setting
the device scale factor to 2.

While we could set the device scale factor to 2 on our workers, I think
that will end up being somewhat disruptive. So, to help make it easier
for people to adopt this change we can handle it at the configuration
level. This will make it so the device scale factor bump comes in with a
pull request, which will avoid having a bunch of confusing diffs for
everyone--since either you will have a merge base before the PR or after
the PR, so the comparisons will consistently use the same device scale
factor.

In order for this to actually work, we will need to deploy a new worker
that is able to consume this value correctly.