Remove waitForSomeContent
This function was added in
https://github.com/happo/happo-plugin-storybook/commit/28c9e16 at a time
when we weren't using Storybook events to know when a story was ready
and rendered. We basically told Storybook to render and then waited up
to 100ms for some markup to show up in the root element.

I'm removing this as part of investigating a strange pattern I'm seeing
in the Happo Storybook test suite for happo.io where all stories take
just over 100ms to render. The closest candidate to adding a 100ms delay
is the waitForSomeContent function, and I feel confident it is not
useful at all since we added the event listener.