Move CLI script to own file
We're seeing people running into issues with the `happo` command
silently exiting directly. This is because `import.meta.main` resolves
to a falsy value, so the main function is never executed.

The check for import.meta.main was added in 9a78605e43c5f93 to make it
easier to test cli/index.ts, conditionally only executing the main
function if it was run as a CLI command. We can keep tests happy by
simply moving the CLI entrypoint out of the way.