Make `happo.init` read stories
I'm fixing the issue where the first example in a storybook build has a
longer render time than it should have. This is because the cost of
reading all stories into memory is paid on the first nextExample call.
I'm changing all this to happen in `init` instead.

There's a corresponding worker change for this as well. It's not
strictly needed but will make us less exposed to non-resolving promises
from `happo.init`:
https://github.com/happo/happo-snap/pull/1032