Use temporary API key/secret locally
If there is no API key or secret configured, we can ask the user for
one. I've worked on a change in happo-view that will show a page with an
"Authenticate" button. This button will cause a short-lived API key to
be created. This key will then be communicated back to the CLI script
using a localhost server and a redirect.