Fix storybook data fetch story test
We were making requests against an external domain that started
responding with 403 forbidden. To fix this, I wanted to spin up little
sever along with storybook to serve some API calls via a vite plugin.
That seemed to work when running Storybook locally, but it did not
resolve the issue when running in Happo. So, instead I decided to use a
different URL for these requests.