Handle 409 status codes from cancel job API endpoint
This happens when the job has already been completed. This is normal
behavior, so we don't want to throw here.

We should possibly also handle 404 this way, which happens when no job
is found for the pair of SHAs. I see that error reported fairly
frequently, but I don't really understand how it comes about yet, so I
don't want to jump into it here just yet.