Throw more helpful error message when config extension is not supported
If people are using an older version of Node.js, it will fail to import
files with some extensions, such as .ts or .mts. The error message is
not very helpful, it just says it is an unknown file extension. I'm
hoping we can help people out a little here by providing a more useful
error message.