Mark modal dialogs with data-happo-modal
We noticed a problem where modal-specific styles that are applied to
modal dialogs (i.e. `:modal`), which are only applied when the dialog is
shown via JS `.showModal()` were not showing up in Happo screenshots.

To make this work, we need the worker to be able to identify which
dialogs were shown via `.showModal()` and which were not. We can achieve
that by adding a data attribute to these dialogs.

This will require a change on the worker before it will fully work.